% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_extension.R
\name{Approx.integration}
\alias{Approx.integration}
\title{Approximate Integration Method Selector}
\usage{
Approx.integration(
  method,
  period,
  y_tilde,
  v_tilde,
  tau_tilde,
  n_samples = NULL,
  n_intervals = NULL,
  n_points = NULL,
  epsilon = NULL
)
}
\arguments{
\item{method}{Character, the integration method to use ("mc", "ti", or "gl").}

\item{period}{Numeric vector, the integration interval.}

\item{y_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{v_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{tau_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{n_samples}{Integer, number of samples for Monte Carlo integration (if `method` is "mc").}

\item{n_intervals}{Integer, number of intervals for Trapezoidal integration (if `method` is "ti").}

\item{n_points}{Integer, number of points for Gaussian-Legendre integration (if `method` is "gl").}

\item{epsilon}{Numeric, small value for Trapezoidal integration (if `method` is "ti").}
}
\value{
The result from the selected integration method.
}
\description{
This function selects the appropriate integration method (Monte Carlo, Trapezoidal, or Gaussian-Legendre)
based on the user's input and calls the corresponding function to compute the integral.
}
