% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{drIG}
\alias{drIG}
\title{Probability Density Function of the Reparameterized Inverse Gaussian (drIG)}
\usage{
drIG(n1, lambda1, gamma1)
}
\arguments{
\item{n1}{Numeric vector of values to evaluate the PDF at.}

\item{lambda1}{Shape parameter.}

\item{gamma1}{Scale parameter.}
}
\value{
A numeric vector of PDF values.
}
\description{
Computes the probability density function (PDF) of the reparameterized inverse Gaussian distribution.
}
