% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_extension.R
\name{f_uz_given_D_gaussian}
\alias{f_uz_given_D_gaussian}
\title{Estimate Function for f(u_ij | D)}
\usage{
f_uz_given_D_gaussian(
  u_ij,
  method = "gl",
  period = c(0, 1),
  y_tilde,
  v_tilde,
  tau_tilde,
  n_samples = NULL,
  n_intervals = NULL,
  n_points = 30
)
}
\arguments{
\item{u_ij}{Numeric, the integration variable.}

\item{method}{Character, the integration method to use ("mc", "ti", "gl").}

\item{period}{Numeric vector, the integration period.}

\item{y_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{v_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{tau_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{n_samples}{Integer, number of samples for Monte Carlo integration.}

\item{n_intervals}{Integer, number of intervals for Trapezoidal integration.}

\item{n_points}{Integer, number of points for Gaussian-Legendre integration.}
}
\value{
The likelihood estimate \( f(u_{ij} | D) \).
}
\description{
This function computes the ratio \( f(u_{ij} | D) \), where \( f(u_{ij} | D) \) is the
likelihood of a given value of \( u_{ij} \) conditioned on the observed data.
}
