% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_path.R
\name{fit.path.plot}
\alias{fit.path.plot}
\title{Plot Fitted Path with or without Confidence Intervals}
\usage{
fit.path.plot(
  dat,
  True_Y,
  t_list,
  u_list,
  select_ID,
  types,
  ci = TRUE,
  scope = NULL
)
}
\arguments{
\item{dat}{A numeric matrix containing the estimated parameters for each product.}

\item{True_Y}{A tibble or data frame containing the true observed values, with columns `TIME`, `UV`, and `observed`.}

\item{t_list}{A list of time values for each product.}

\item{u_list}{A list of covariate values for each product (can be NULL).}

\item{select_ID}{A vector of product IDs to plot.}

\item{types}{A vector specifying the type of model used for the Lambda calculation.}

\item{ci}{A logical value indicating whether to include confidence intervals in the plot (default is TRUE).}

\item{scope}{A numeric vector of length 2 specifying the quantile range for the confidence intervals (e.g., `c(0.025, 0.975)`).}
}
\value{
A ggplot object representing the fitted degradation paths for the selected product units.
}
\description{
This function generates a plot of the fitted degradation paths for each unit, with an option to display confidence
intervals based on quantiles. The function supports models "M0", "M1", "M2" and allows for customized plotting.
}
