% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_path.R
\name{fit.path.process}
\alias{fit.path.process}
\title{Compute Fitted Path for Each Product}
\usage{
fit.path.process(dat, select_ID, t_list, u_list, types)
}
\arguments{
\item{dat}{A numeric vector of estimated parameters (including Lambda values) for the product.}

\item{select_ID}{A vector of product IDs to select for the fitted path calculation.}

\item{t_list}{A list where each element contains the time points for each product.}

\item{u_list}{A list where each element contains the covariates for each product.}

\item{types}{A vector specifying the type of model used for the Lambda calculation.}
}
\value{
A tibble containing the fitted degradation values (`Y`) for each product unit over time.
}
\description{
This function calculates the fitted degradation values for each product unit based on estimated parameters
using the Lambda function. The function assumes two scales, computing the cumulative fitted degradation values
for each unit.
}
