% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_extension.R
\name{gaussian_legendre_integral}
\alias{gaussian_legendre_integral}
\title{Gaussian-Legendre Integration}
\usage{
gaussian_legendre_integral(
  period = c(0, 1),
  y_tilde,
  v_tilde,
  tau_tilde,
  n_points = 10
)
}
\arguments{
\item{period}{Numeric vector, the integration interval, default is `c(0, 1)`.}

\item{y_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{v_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{tau_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{n_points}{Integer, the number of Gaussian points to use for the integration.}
}
\value{
A list containing:
  - `integral_estimate`: Estimated integral using the Gaussian-Legendre quadrature.
  - `time_avg`: Time taken to perform the calculation.
}
\description{
This function estimates the integral using the Gaussian-Legendre quadrature method.
It uses Legendre polynomial roots and weights to compute the integral over the given interval.
}
