% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_guess.R
\name{init.log.likelihood}
\alias{init.log.likelihood}
\title{Initial Log-Likelihood Calculation}
\usage{
init.log.likelihood(
  params = rep(1, 5),
  model,
  type1,
  data1 = sim_dat$diff_Y_t[, 1],
  t = t,
  u = NULL
)
}
\arguments{
\item{params}{A numeric vector containing initial values for the parameters.
The vector should be of length 5 for models "M0", "M1", "M2", or length 3 for models "M3" and "M4".}

\item{model}{A character string specifying the model type. Must be one of "M0", "M1", "M2", "M3", or "M4".}

\item{type1}{A string specifying the type of distribution or model for the lambda calculation.}

\item{data1}{A numeric vector containing the observed data for a single product (response variable).}

\item{t}{A numeric vector of time or other covariates used in the lambda function.}

\item{u}{A numeric vector of additional covariates for the lambda function (can be NULL for some models).}
}
\value{
A numeric value representing the negative log-likelihood for the given parameters.
}
\description{
This function computes the log-likelihood for a given set of parameters for different models.
The models "M0", "M1", "M2" involve gamma and lambda terms, while models "M3" and "M4" involve
only lambda. The function calculates the log-likelihood based on the observed data.
}
