% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_extension.R
\name{Integrand_fun}
\alias{Integrand_fun}
\title{Integrand Function for the Gaussian-Legendre Integral}
\usage{
Integrand_fun(u_ij, y_tilde, v_tilde, tau_tilde)
}
\arguments{
\item{u_ij}{Numeric, a sample value for the integration variable.}

\item{y_tilde}{Numeric, parameter for the exponential function.}

\item{v_tilde}{Numeric, parameter for the exponential function.}

\item{tau_tilde}{Numeric, parameter for the exponential function.}
}
\value{
The value of the integrand at the given `u_ij`.
}
\description{
This function defines the integrand for the Gaussian-Legendre integration method.
It calculates the product of two parts:
1. A power function of \( u_{ij} \), and
2. An exponential function involving `y_tilde`, `v_tilde`, and `tau_tilde`.
}
