% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{Lambda_fun_der}
\alias{Lambda_fun_der}
\title{Lambda Function Derivative Calculation}
\usage{
Lambda_fun_der(type, model, par0, t, u)
}
\arguments{
\item{type}{The model type (e.g., "M0").}

\item{model}{The specific model for differentiation.}

\item{par0}{A vector of model parameters.}

\item{t}{Time vector.}

\item{u}{Another variable (e.g., cycles or units).}
}
\value{
A list containing derivatives of Lambda functions for both time and another variable.
}
\description{
Computes the derivatives of Lambda functions with respect to model parameters.
}
