% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_extension.R
\name{mc_integral}
\alias{mc_integral}
\title{Monte Carlo Integration}
\usage{
mc_integral(period = c(0, 1), n_samples, y_tilde, v_tilde, tau_tilde)
}
\arguments{
\item{period}{Numeric vector, the integration interval, default is `c(0, 1)`.}

\item{n_samples}{Integer, number of samples to use for the Monte Carlo integration.}

\item{y_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{v_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{tau_tilde}{Numeric, parameter for the exponential function in the integrand.}
}
\value{
A list containing:
  - `integral_estimate`: Estimated integral using Monte Carlo.
  - `time_avg`: Time taken to perform the calculation.
}
\description{
This function estimates an integral using the Monte Carlo method. It randomly samples
`u_ij` values from a uniform distribution and computes the mean of the integrand
values for these samples.
}
