% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_fct.R
\name{move_infinite}
\alias{move_infinite}
\title{Function to Remove Rows Containing Infinite Values or Exceeding Thresholds}
\usage{
move_infinite(df)
}
\arguments{
\item{df}{A data frame or matrix where rows are being filtered.}
}
\value{
A data frame or matrix with rows that meet the conditions removed.
}
\description{
This function applies the `selected_fun` to each row of the data frame to filter out rows
that contain infinite values or exceed a threshold. It is used to clean the data before
further analysis.
}
