% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_fct.R
\name{performence.compare}
\alias{performence.compare}
\title{Function to Compute Performance Metrics (RB, MSE, RRMSE, CP, LEN)}
\usage{
performence.compare(
  true_value = para,
  final_par = c(em_para_re, inti_par, bayes_para_re),
  interval_re = NULL
)
}
\arguments{
\item{true_value}{A numeric vector of true parameter values.}

\item{final_par}{A numeric vector or matrix containing the estimated parameters.}

\item{interval_re}{A data frame containing the confidence interval bounds (Lower, Upper) for the parameters.
Default is NULL.}
}
\value{
A list containing the computed values for RB, MSE, re_MSE, CP, and LEN.
}
\description{
This function computes several performance metrics comparing the estimated parameters (`final_par`)
to the true parameter values (`true_value`). It calculates relative bias (RB), mean squared error (MSE),
relative mean squared error (re_MSE), coverage probability (CP), and length of the confidence interval (LEN).
}
