% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{prIG}
\alias{prIG}
\title{Cumulative Distribution Function of the Reparameterized Inverse Gaussian (prIG)}
\usage{
prIG(p, lambda1, gamma1)
}
\arguments{
\item{p}{Numeric vector of probabilities to evaluate the CDF at.}

\item{lambda1}{Shape parameter.}

\item{gamma1}{Scale parameter.}
}
\value{
A numeric vector of CDF values.
}
\description{
Computes the cumulative distribution function (CDF) of the reparameterized inverse Gaussian distribution.
}
