% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{Reliability}
\alias{Reliability}
\title{Compute Reliability over Time}
\usage{
Reliability(
  par = em_para_re[[1]],
  model = models,
  t,
  u,
  type = types,
  D = 10,
  bin = 10
)
}
\arguments{
\item{par}{A numeric vector of parameters used in the model.}

\item{model}{A character string indicating the model type ("M0", "M1", "M2", "M3", "M4").}

\item{t}{A numeric vector of time points at which to evaluate the reliability.}

\item{u}{A numeric vector of uncertainty or additional variables, default is NULL.}

\item{type}{A character string indicating the type of distribution or function to use (e.g., "Gamma", "Weibull").}

\item{D}{A numeric value indicating a scale parameter, default is 10.}

\item{bin}{A numeric value indicating the number of bins to use when discretizing the time range, default is 10.}
}
\value{
A list containing:
\item{result}{A data frame with the computed reliability for each time point.}
\item{MTTF}{The computed Mean Time to Failure (MTTF).}
\item{plot}{A ggplot object showing the reliability over time.}
}
\description{
This function calculates the reliability function over a range of time points for a given model and parameters.
It can generate a plot of the reliability function and compute the Mean Time to Failure (MTTF) based on the calculated CDF.
}
