% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{rrIG}
\alias{rrIG}
\title{Reparameterized Inverse Gaussian Random Number Generator (rIG)}
\usage{
rrIG(n1, lambda1, gamma1)
}
\arguments{
\item{n1}{Number of random samples.}

\item{lambda1}{Shape parameter of the distribution.}

\item{gamma1}{Scale parameter of the distribution.}
}
\value{
A numeric vector of random samples.
}
\description{
Generates random samples from a reparameterized inverse Gaussian distribution.
}
