% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_fct.R
\name{save.result}
\alias{save.result}
\title{Function to Save Simulation Results}
\usage{
save.result(
  mttf_re = mttf_re,
  sim_re = sim_re,
  MTTF = TRUE,
  col_names = c("alpha_t", "beta_t", "alpha_u", "beta_u", "kappa", "sigma2")
)
}
\arguments{
\item{mttf_re}{A list containing results from MTTF simulations.}

\item{sim_re}{A list containing simulation results with keys RB, MSE, re_MSE, etc.}

\item{MTTF}{Logical, indicating whether to compute MTTF statistics. Default is TRUE.}

\item{col_names}{A character vector containing column names for the results. Default includes
"alpha_t", "beta_t", "alpha_u", "beta_u", "kappa", and "sigma2".}
}
\value{
A list of data frames containing the calculated statistics.
}
\description{
This function processes and saves the results of a simulation, calculating summary statistics
such as relative bias (RB), root mean squared error (RMSE), relative RMSE (RRMSE),
coverage probability (CP), and confidence interval length (LEN). The results are saved into
a list of data frames corresponding to different statistics.
}
