% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_fct.R
\name{selected_fun}
\alias{selected_fun}
\title{Function to Filter Rows Based on Criteria}
\usage{
selected_fun(row)
}
\arguments{
\item{row}{A numeric vector representing a single row of data.}
}
\value{
A logical value: TRUE if the row should be kept, FALSE otherwise.
}
\description{
This function is used to filter out rows that contain infinite values or values greater than a threshold.
It checks whether any of the columns are set to FALSE for EM_scen, Bayes_scen, or interval_scen,
and also whether the row contains any infinite values or values exceeding 1000.
}
