% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{sim.dat.path}
\alias{sim.dat.path}
\title{Simulate Degradation Data}
\usage{
sim.dat.path(model = "M0", type, m, n, t, u, par, gamma = 1)
}
\arguments{
\item{model}{A character string specifying the model type (e.g., "M0", "M1", etc.).}

\item{type}{The type of Lambda function (e.g., "pp").}

\item{m}{The number of steps or units.}

\item{n}{The number of samples or simulations.}

\item{t}{Time vector.}

\item{u}{A second variable (optional).}

\item{par}{A vector of parameters for the model.}

\item{gamma}{Scale parameter (optional, defaults to 1).}
}
\value{
A list containing the simulated degradation paths for different models.
}
\description{
Simulates the degradation path data based on the specified model and parameters.
}
