% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_extension.R
\name{trapezoidal_integral}
\alias{trapezoidal_integral}
\title{Trapezoidal Rule Integration}
\usage{
trapezoidal_integral(
  period = c(0, 1),
  y_tilde,
  v_tilde,
  tau_tilde,
  n_intervals = 1000,
  epsilon = 1e-06
)
}
\arguments{
\item{period}{Numeric vector, the integration interval, default is `c(0, 1)`.}

\item{y_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{v_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{tau_tilde}{Numeric, parameter for the exponential function in the integrand.}

\item{n_intervals}{Integer, the number of intervals for the trapezoidal rule.}

\item{epsilon}{Numeric, small value to avoid sampling at the boundaries (0 or 1).}
}
\value{
A list containing:
  - `integral_estimate`: Estimated integral using the trapezoidal rule.
  - `time_avg`: Time taken to perform the calculation.
}
\description{
This function estimates the integral using the trapezoidal rule. It creates a set of evenly spaced
`u_ij` values and computes the average of the integrand at the boundaries and at the intermediate points.
}
