% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{uncon_lifetime_CDF}
\alias{uncon_lifetime_CDF}
\title{Reliability Function Calculation}
\usage{
uncon_lifetime_CDF(model, par = em_para_re, type, t_ind, u_ind, D = 300)
}
\arguments{
\item{model}{A character string indicating the model type. It can be one of "M0", "M1", "M2", "M3", "M4".}

\item{par}{A numeric vector of parameters used in the model. The length and meaning of the vector depend on the selected model.}

\item{type}{A character string indicating the type of distribution or function to use (e.g., "Gamma", "Weibull").}

\item{t_ind}{A numeric value indicating the time at which to evaluate the CDF.}

\item{u_ind}{A numeric value indicating the additional variable (e.g., uncertainty, user-defined parameter).}

\item{D}{A numeric value indicating a scale parameter, default is 300.}
}
\value{
A numeric value representing the reliability at time `t_ind` and additional parameters.
}
\description{
This function computes the reliability function (CDF) based on a given model and parameters.
It is designed to handle multiple models (M0, M1, M2, M3, M4) and computes the reliability over a specified range of time.
}
